requirements.txt
docopt
gevent
websocket-client
